pyproject.toml
[tool.poetry]
name = "sarjana"
version = "0.1.3"
description = ""
authors = ["Thaza_Kun <61819672+Thaza-Kun@users.noreply.github.com>"]
packages = [
    { include = "sarjana" }
]

[tool.poetry.dependencies]
python = "^3.9"
asyncio = "^3.4.3"
mw-plot = "^0.9.0"
umap-learn = "^0.5.3"
seaborn = "^0.12.1"
pandas = "^1.5.1"
numpy = "^1.23.4"
jupyter = "^1.0.0"
hdbscan = "^0.8.29"
plotly = "^5.11.0"
jupytext = "^1.14.1"

[tool.poetry.dev-dependencies]
nbformat = "^5.7.0"
jupyterlab = "^3.5.0"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
